Move logic for flag conversion into constructor
